INSIGHTSAPI/requirements.txt
celery==5.4.0
chardet==5.2.0
Django==5.0.7
django_auth_ldap==4.8.0
django-cors-headers==4.4.0
django_sendfile2==0.7.1
django_simple_history==3.7.0
djangorestframework_simplejwt==5.3.1
drf_spectacular==0.27.2
factory_boy==3.3.0
Faker==26.0.0
freezegun==1.5.1
ftfy==6.2.0
holidays==0.53
mysqlclient==2.2.4
mysql-connector-python==9.0.0
num2words==0.5.13
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
pdfkit==1.0.0
Pillow==10.4.0
python-dotenv==1.0.1
python_ldap==3.4.4
python_magic==0.4.27
Requests==2.32.3
redis==5.0.8
gunicorn==22.0.0
locust==2.31.1
django-redis==5.4.0
django-debug-toolbar==4.4.6
django-stubs==5.1.0
djangorestframework-stubs==3.15.1